requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
Langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf

